Remove unnecessary uiReducer
The uiReducer is now replaced with a nested combineReducers
in the rootReducer. Makes the sub reducer initial state
easy to define and easier to test.
Refs #84.
