+++ app/reducers/index.js
@@ -4,14 +4,20 @@ import { routerStateReducer } from 'redux-router';
 import apiReducer from 'reducers/apiReducer';
 import authReducer from 'reducers/authReducer';
 import dataReducer from 'reducers/dataReducer';
-import uiReducer from 'reducers/uiReducer';
+import modalsReducer from 'reducers/modalsReducer';
+import reservationReducer from 'reducers/reservationReducer';
+import searchReducer from 'reducers/searchReducer';
 
 const rootReducer = combineReducers({
   api: apiReducer,
   auth: authReducer,
   data: dataReducer,
   router: routerStateReducer,
-  ui: uiReducer,
+  ui: combineReducers({
+    modals: modalsReducer,
+    reservation: reservationReducer,
+    search: searchReducer,
+  }),
 });
 
 export default rootReducer;
